Struct ultraviolet::bivec::Bivec2x4

#[repr(C)]
pub struct Bivec2x4 {
    pub xy: f32x4,
}

A bivector in 2d space.

Since in 2d there is only one plane in the whole of 2d space, a 2d bivector has only one component.

Please see the module level documentation for more information on bivectors generally!

Fields
xy: f32x4
Implementations

impl Bivec2x4

pub const fn new(xy: f32x4) -> Self

pub fn zero() -> Self

pub fn unit_xy() -> Self

pub fn mag_sq(&self) -> f32x4

pub fn mag(&self) -> f32x4

pub fn normalize(&mut self)

pub fn normalized(&self) -> Self

pub fn dot(&self, rhs: Self) -> f32x4

pub fn layout() -> Layout

pub fn as_slice(&self) -> &[f32x4]

pub fn as_byte_slice(&self) -> &[u8]

pub fn as_mut_slice(&mut self) -> &mut [f32x4]

pub fn as_mut_byte_slice(&mut self) -> &mut [u8]

pub const fn as_ptr(&self) -> *const f32x4

Returns a constant unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

pub fn as_mut_ptr(&mut self) -> *mut f32x4

Returns a mutable unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.