Struct ultraviolet::bivec::WBivec3

#[repr(C)]pub struct WBivec3 {
    pub xy: f32x4,
    pub xz: f32x4,
    pub yz: f32x4,
}

A bivector in 3d space.

In 3d, a bivector has 3 components, each one representing the signed projected area of the bivector onto one of the 3 basis bivectors, which can be thought of as corresponding to each of hte three basis planes. This is analogous to the components of a 3d vector, which correspond to the projected length of the vector onto the three basis *vectors. Since in 3d, there are three components for both vectors and bivectors, 3d bivectors have been historically confused with 3d vectors quite a lot.

Please see the module level documentation for more information on bivectors generally!

Fields
xy: f32x4xz: f32x4yz: f32x4
Implementations

impl WBivec3

pub fn new(xy: f32x4, xz: f32x4, yz: f32x4) -> Self

pub fn zero() -> Self

pub fn from_normalized_axis(v: Wec3) -> Self

Create the bivector which represents the same plane of rotation as a given normalized 'axis vector'

pub fn unit_xy() -> Self

pub fn unit_xz() -> Self

pub fn unit_yz() -> Self

pub fn mag_sq(&self) -> f32x4

pub fn mag(&self) -> f32x4

pub fn normalize(&mut self)

pub fn normalized(&self) -> Self

pub fn layout() -> Layout

pub fn as_slice(&self) -> &[f32x4]

pub fn as_byte_slice(&self) -> &[u8]

pub fn as_mut_slice(&mut self) -> &mut [f32x4]

pub fn as_mut_byte_slice(&mut self) -> &mut [u8]

pub fn as_ptr(&self) -> *const f32x4

Returns a constant unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

pub fn as_mut_ptr(&mut self) -> *mut f32x4

Returns a mutable unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.