Struct ultraviolet::int::Vec4u

#[repr(C)]pub struct Vec4u {
    pub x: u32,
    pub y: u32,
    pub z: u32,
    pub w: u32,
}

A set of four coordinates which may be interpreted as a point or vector in 4d space, or as a homogeneous 3d vector or point.

Generally this distinction between a point and vector is more of a pain than it is worth to distinguish on a type level, however when converting to and from homogeneous coordinates it is quite important.

Fields
x: u32y: u32z: u32w: u32
Implementations

impl Vec4u

pub fn new<T: Into<u32>>(x: T, y: T, z: T, w: T) -> Self

pub fn broadcast<T: Into<u32> + Copy>(val: T) -> Self

pub fn unit_x() -> Self

pub fn unit_y() -> Self

pub fn unit_z() -> Self

pub fn unit_w() -> Self

pub fn dot(&self, other: Vec4u) -> u32

pub fn reflect(&mut self, normal: Vec4u)

pub fn reflected(&self, normal: Vec4u) -> Self

pub fn mag(&self) -> u32

pub fn mag_sq(&self) -> u32

pub fn mul_add(&self, mul: Vec4u, add: Vec4u) -> Self

pub fn clamp(&mut self, min: Self, max: Self)

pub fn clamped(self, min: Self, max: Self) -> Self

pub fn map<F>(&self, f: F) -> Self where
F: Fn(u32) -> u32,

pub fn apply<F>(&mut self, f: F) where
F: Fn(u32) -> u32,

pub fn max_by_component(self, other: Self) -> Self

pub fn min_by_component(self, other: Self) -> Self

pub fn component_max(&self) -> u32

pub fn component_min(&self) -> u32

pub fn zero() -> Self

pub fn one() -> Self

pub fn xy(&self) -> Vec2u

pub fn xyz(&self) -> Vec3u

pub fn layout() -> Layout

pub fn as_slice(&self) -> &[u32]

pub fn as_array(&self) -> [u32; 4]

pub fn as_byte_slice(&self) -> &[u8]

pub fn as_mut_slice(&mut self) -> &mut [u32]

pub fn as_mut_byte_slice(&mut self) -> &mut [u8]

pub fn as_ptr(&self) -> *const u32

Returns a constant unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

pub fn as_mut_ptr(&mut self) -> *mut u32

Returns a mutable unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.