Trait ultraviolet::interp::Slerp

pub trait Slerp<T> {
    fn slerp(&self, end: Self, t: T) -> Self;
}

Spherical-linear interpolation.

Basically, interpolation that maintains a constant angular velocity from one orientation on a unit hypersphere to another. This is sorta the “high quality” interpolation for Rotors, and it can also be used to interpolate other things, one example being interpolation of 3d normal vectors.

Note that you should often normalize the result returned by this operation, when working with Rotors, etc!

Required methods

fn slerp(&self, end: Self, t: T) -> Self