Struct ultraviolet::mat::Mat2x4

#[repr(C)]
pub struct Mat2x4 {
    pub cols: [Vec2x4; 2],
}

A 2x2 square matrix.

Useful for performing linear transformations (rotation, scaling) on 2d vectors.

Fields
cols: [Vec2x4; 2]
Implementations

impl Mat2x4

pub const fn new(col1: Vec2x4, col2: Vec2x4) -> Self

pub fn identity() -> Self

pub fn into_homogeneous(self) -> Mat3x4

Turn this into a homogeneous 2d transformation matrix.

pub fn transpose(&mut self)

pub fn transposed(&self) -> Self

pub fn determinant(&self) -> f32x4

pub fn adjugate(&self) -> Self

The adjugate of this matrix, i.e. the transpose of the cofactor matrix.

This is equivalent to the inverse but without dividing by the determinant of the matrix, which can be useful in some contexts for better performance.

One such case is when this matrix is interpreted as a a homogeneous transformation matrix, in which case uniform scaling will not affect the resulting projected 3d version of transformed points or vectors.

pub fn inverse(&mut self)

If this matrix is not currently invertable, this function will return an invalid inverse. This status is not checked by the library.

pub fn inversed(&self) -> Self

If this matrix is not currently invertable, this function will return an invalid inverse. This status is not checked by the library.

pub fn layout() -> Layout

pub fn as_array(&self) -> &[f32x4; 4]

pub fn as_component_array(&self) -> &[Vec2x4; 2]

pub fn as_slice(&self) -> &[f32x4]

pub fn as_component_slice(&self) -> &[Vec2x4]

pub fn as_byte_slice(&self) -> &[u8]

pub fn as_mut_slice(&mut self) -> &mut [f32x4]

pub fn as_mut_component_slice(&mut self) -> &mut [Vec2x4]

pub fn as_mut_byte_slice(&mut self) -> &mut [u8]

pub const fn as_ptr(&self) -> *const f32x4

Returns a constant unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

pub fn as_mut_ptr(&mut self) -> *mut f32x4

Returns a mutable unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.