Struct ultraviolet::mat::Mat3

#[repr(C)]pub struct Mat3 {
    pub cols: [Vec3; 3],
}

A 3x3 square matrix.

Useful for performing linear transformations (rotation, scaling) on 3d vectors, or for performing arbitrary transformations (linear + translation, projection, etc) on homogeneous 2d vectors

Fields
cols: [Vec3; 3]
Implementations

impl Mat3

pub fn new(col1: Vec3, col2: Vec3, col3: Vec3) -> Self

pub fn from_translation(trans: Vec2) -> Self

Assumes homogeneous 2d coordinates.

pub fn from_scale_homogeneous(scale: f32) -> Self

Assumes homogeneous 2d coordinates.

pub fn from_nonuniform_scale_homogeneous(scale: Vec3) -> Self

Assumes homogeneous 2d coordinates.

pub fn from_rotation_homogeneous(angle: f32) -> Self

Builds a homogeneous 2d rotation matrix (in the xy plane) from a given angle in radians.

pub fn from_scale(scale: f32) -> Self

pub fn from_nonuniform_scale(scale: Vec3) -> Self

pub fn identity() -> Self

pub fn from_euler_angles(roll: f32, pitch: f32, yaw: f32) -> Self

Angles are applied in the order roll -> pitch -> yaw.

Yaw is rotation inside the xz plane ("around the y axis")

Pitch is rotation inside the yz plane ("around the x axis")

Roll is rotation inside the xy plane ("around the z axis")

Important: This function assumes a right-handed, y-up coordinate space where:

+X axis points right

+Y axis points up

+Z axis points towards the viewer (i.e. out of the screen)

This means that you may see unexpected behavior when used with OpenGL or DirectX as they use a different coordinate system. You should use the appropriate projection matrix in projection module to fit your use case to remedy this.

pub fn from_rotation_x(angle: f32) -> Self

Create a new rotation matrix from a rotation "about the x axis". This is here as a convenience function for users coming from other libraries; it is more proper to think of this as a rotation in the yz plane.

Important: This function assumes a right-handed, y-up coordinate space where:

+X axis points right

+Y axis points up

+Z axis points towards the viewer (i.e. out of the screen)

This means that you may see unexpected behavior when used with OpenGL or DirectX as they use a different coordinate system. You should use the appropriate projection matrix in projection module to fit your use case to remedy this.

pub fn from_rotation_y(angle: f32) -> Self

Create a new rotation matrix from a rotation "about the y axis". This is here as a convenience function for users coming from other libraries; it is more proper to think of this as a rotation in the xz plane.

Important: This function assumes a right-handed, y-up coordinate space where:

+X axis points right

+Y axis points up

+Z axis points towards the viewer (i.e. out of the screen)

This means that you may see unexpected behavior when used with OpenGL or DirectX as they use a different coordinate system. You should use the appropriate projection matrix in projection module to fit your use case to remedy this.

pub fn from_rotation_z(angle: f32) -> Self

Create a new rotation matrix from a rotation "about the z axis". This is here as a convenience function for users coming from other libraries; it is more proper to think of this as a rotation in the xy plane.

Important: This function assumes a right-handed, y-up coordinate space where:

+X axis points right

+Y axis points up

+Z axis points towards the viewer (i.e. out of the screen)

This means that you may see unexpected behavior when used with OpenGL or DirectX as they use a different coordinate system. You should use the appropriate projection matrix in projection module to fit your use case to remedy this.

pub fn from_angle_plane(angle: f32, plane: Bivec3) -> Self

Construct a rotation matrix given a bivector which defines a plane and rotation orientation, and a rotation angle.

plane must be normalized!

This is the equivalent of an axis-angle rotation.

pub fn into_homogeneous(self) -> Mat4

pub fn inverse(&mut self)

If this matrix is not currently invertable, this function will return an invalid inverse. This status is not checked by the library.

pub fn inversed(&self) -> Self

If this matrix is not currently invertable, this function will return an invalid inverse. This status is not checked by the library.

pub fn transpose(&mut self)

pub fn transposed(&self) -> Self

pub fn layout() -> Layout

pub fn as_array(&self) -> &[f32; 9]

pub fn as_component_array(&self) -> &[Vec3; 3]

pub fn as_slice(&self) -> &[f32]

pub fn as_component_slice(&self) -> &[Vec3]

pub fn as_byte_slice(&self) -> &[u8]

pub fn as_mut_slice(&mut self) -> &mut [f32]

pub fn as_mut_component_slice(&mut self) -> &mut [Vec3]

pub fn as_mut_byte_slice(&mut self) -> &mut [u8]

pub fn as_ptr(&self) -> *const f32

Returns a constant unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

pub fn as_mut_ptr(&mut self) -> *mut f32

Returns a mutable unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.