Module ultraviolet::projection

Utility functions to create projection matrices.

Each module contains versions for a certain kind of “base” coordinate systems, i.e. the coordinate system that your client application is using. All of them assume +x is pointing to the right.

The rh_yup module is publicly re-exported as it contains the functions for what is generally seen as the ‘standard’ computer graphics coordinate space.

Re-exports

pub use rh_yup::*;

Modules

lh_ydown

Projection matrices that are intended to be used when the base coordinate system (i.e. the one used by the application code) is left-handed, with the x-axis pointing right, y-axis pointing down, and z-axis pointing out of the screen.

lh_yup

Projection matrices that are intended to be used when the base coordinate system (i.e. the one used by the application code) is left-handed, with the x-axis pointing right, y-axis pointing up, and z-axis pointing into the screen.

rh_ydown

Projection matrices that are intended to be used when the base coordinate system (i.e. the one used by the application code) is right-handed, with the x-axis pointing right, y-axis pointing down, and z-axis pointing into the screen.

rh_yup

Projection matrices that are intended to be used when the base coordinate system (i.e. the one used by the application code) is right-handed with the the x-axis pointing right, y-axis pointing up, and z-axis pointing out of the screen. This is reexported at the root of projections as it is the de-facto standard coordinate system for doing computer graphics programming.