Module ultraviolet::projection::lh_ydown

Projection matrices that are intended to be used when the base coordinate system (i.e. the one used by the application code) is left-handed, with the x-axis pointing right, y-axis pointing down, and z-axis pointing out of the screen.

Functions

orthographic_gl

Orthographic projection matrix for use with OpenGL.

orthographic_vk

Orthographic projection matrix for use with Vulkan.

orthographic_wgpu_dx

Orthographic projection matrix for use with WebGPU or DirectX.

perspective_gl

Perspective projection matrix meant to be used with OpenGL.

perspective_infinite_z_gl

Perspective projection matrix with infinite z-far plane meant to be used with OpenGL.

perspective_infinite_z_vk

Perspective projection matrix with infinite z-far plane meant to be used with Vulkan.

perspective_infinite_z_wgpu_dx

Perspective projection matrix with infinite z-far plane meant to be used with WebGPU or DirectX.

perspective_reversed_infinite_z_vk

Perspective projection matrix with reversed and infinite z-axis meant to be used with Vulkan.

perspective_reversed_infinite_z_wgpu_dx_gl

Perspective projection matrix with reversed and infinite z-axis meant to be used with WebGPU, OpenGL or DirectX.

perspective_reversed_z_vk

Perspective projection matrix with reversed z-axis meant to be used with Vulkan.

perspective_reversed_z_wgpu_dx_gl

Perspective projection matrix with reversed z-axis meant to be used with WebGPU, DirectX, or OpenGL.

perspective_vk

Perspective projection matrix meant to be used with Vulkan.

perspective_wgpu_dx

Perspective projection matrix meant to be used with WebGPU or DirectX.