Module ultraviolet::projection::rh_ydown

Projection matrices that are intended to be used when the base coordinate system (i.e. the one used by the application code) is right-handed, with the x-axis pointing right, y-axis pointing down, and z-axis pointing into the screen.

Functions

orthographic_dx
Orthographic projection matrix for use with OpenGL.

orthographic_gl
Orthographic projection matrix for use with OpenGL.

orthographic_vk
Orthographic projection matrix for use with OpenGL.

perspective_dx
Perspective projection matrix meant to be used with DirectX.

perspective_gl
Perspective projection matrix meant to be used with OpenGL.

perspective_infinite_z_dx
Perspective projection matrix with infinite z-far plane meant to be used with DirectX.

perspective_infinite_z_gl
Perspective projection matrix with infinite z-far plane meant to be used with OpenGL.

perspective_infinite_z_vk
Perspective projection matrix with infinite z-far plane meant to be used with Vulkan.

perspective_reversed_infinite_z_dx_gl
Perspective projection matrix with reversed and infinite z-axis meant to be used with OpenGL or DirectX.

perspective_reversed_infinite_z_vk
Perspective projection matrix with reversed and infinite z-axis meant to be used with Vulkan.

perspective_reversed_z_dx_gl
Perspective projection matrix with reversed z-axis meant to be used with DirectX or OpenGL.

perspective_reversed_z_vk
Perspective projection matrix with reversed z-axis meant to be used with Vulkan.

perspective_vk
Perspective projection matrix meant to be used with Vulkan.