Struct ultraviolet::rotor::Rotor2

#[repr(C)]pub struct Rotor2 {
    pub s: f32,
    pub bv: Bivec2,
}

A Rotor in 2d space.

Please see the module level documentation for more information on rotors!

Fields
s: f32bv: Bivec2
Implementations

impl Rotor2

pub fn new(scalar: f32, bivector: Bivec2) -> Self

pub fn identity() -> Self

pub fn from_rotation_between(from: Vec2, to: Vec2) -> Self

Construct a Rotor that rotates one vector to another.

pub fn from_angle_plane(planeangle: Bivec2) -> Self

Construct a rotor given a bivector which defines a plane, rotation orientation, and rotation angle. The bivector defines the plane and orientation, and its magnitude defines the angle of rotation in radians. In 2d, there is only one possible plane of rotation, but two possible orientations of rotation in that plane.

pub fn from_angle(angle: f32) -> Self

Construct a rotor given only an angle. This is possible in 2d since there is only one possible plane of rotation. However, there are two possible orientations. This function uses the common definition of positive angle in 2d as meaning the direction which brings the x unit vector towards the y unit vector.

pub fn mag_sq(&self) -> f32

pub fn mag(&self) -> f32

pub fn normalize(&mut self)

pub fn normalized(&self) -> Self

pub fn reverse(&mut self)

pub fn reversed(&self) -> Self

pub fn rotate_by(&mut self, other: Self)

Rotates this rotor by another rotor in-place. Note that if you are looking to compose rotations, you should NOT use this operation and rather just use regular left-multiplication like for matrix composition.

pub fn rotated_by(self, other: Self) -> Self

Rotates this rotor by another rotor and returns the result. Note that if you are looking to compose rotations, you should NOT use this operation and rather just use regular left-multiplication like for matrix composition.

pub fn rotate_vec(self, vec: &mut Vec2)

Rotates a vector by this rotor.

self must be normalized!

pub fn into_matrix(self) -> Mat2

pub fn layout() -> Layout