Struct ultraviolet::rotor::Rotor2x4

#[repr(C)]
pub struct Rotor2x4 {
    pub s: f32x4,
    pub bv: Bivec2x4,
}

A Rotor in 2d space.

Please see the module level documentation for more information on rotors!

Fields
s: f32x4bv: Bivec2x4
Implementations

impl Rotor2x4

pub const fn new(scalar: f32x4, bivector: Bivec2x4) -> Self

pub fn identity() -> Self

pub fn from_rotation_between(from: Vec2x4, to: Vec2x4) -> Self

Construct a Rotor that rotates one vector to another.

A rotation between antiparallel vectors is undefined!

pub fn from_angle_plane(angle: f32x4, plane: Bivec2x4) -> Self

Construct a rotor given a bivector which defines a plane and rotation orientation, and a rotation angle.

plane must be normalized!

This is the equivalent of an axis-angle rotation.

pub fn from_angle(angle: f32x4) -> Self

Construct a rotor given only an angle. This is possible in 2d since there is only one possible plane of rotation. However, there are two possible orientations. This function uses the common definition of positive angle in 2d as meaning the direction which brings the x unit vector towards the y unit vector.

pub fn mag_sq(&self) -> f32x4

pub fn mag(&self) -> f32x4

pub fn normalize(&mut self)

pub fn normalized(&self) -> Self

pub fn reverse(&mut self)

pub fn reversed(&self) -> Self

pub fn dot(&self, rhs: Self) -> f32x4

pub fn rotate_by(&mut self, other: Self)

Rotates this rotor by another rotor in-place. Note that if you are looking to compose rotations, you should NOT use this operation and rather just use regular left-multiplication like for matrix composition.

pub fn rotated_by(self, other: Self) -> Self

Rotates this rotor by another rotor and returns the result. Note that if you are looking to compose rotations, you should NOT use this operation and rather just use regular left-multiplication like for matrix composition.

pub fn rotate_vec(self, vec: &mut Vec2x4)

Rotates a vector by this rotor.

self must be normalized!

pub fn into_matrix(self) -> Mat2x4

pub fn layout() -> Layout