Struct ultraviolet::transform::Similarity2

#[repr(C)]pub struct Similarity2 {
    pub translation: Vec2,
    pub rotation: Rotor2,
    pub scale: f32,
}

A Similarity, i.e. an Isometry but with an added uniform scaling.

Defined as a uniform scaling followed by a rotation followed by a translation.

You may want to us this type over the corresponding type of homogeneous transformation matrix because it will be faster in most operations, especially composition and inverse.

Fields
translation: Vec2rotation: Rotor2scale: f32
Implementations

impl Similarity2

pub fn new(translation: Vec2, rotation: Rotor2, scale: f32) -> Self

pub fn identity() -> Self

pub fn prepend_scaling(&mut self, scaling: f32)

Add a scaling before this similarity.

This means the scaling will only affect the scaling part of this similarity, not the translational part.

pub fn append_scaling(&mut self, scaling: f32)

Add a scaling after this similarity.

This means the scaling will affect both the scaling and translational parts of this similairty, since it is being applied after this similarity's translational part.

pub fn prepend_rotation(&mut self, rotor: Rotor2)

Add a rotation before this similarity.

This means the rotation will only affect the rotational part of this similarity, not the translational part.

pub fn append_rotation(&mut self, rotor: Rotor2)

Add a rotation after this similarity.

This means the rotation will affect both the rotational and translational parts of this similarity, since it is being applied after this similarity's translational part.

pub fn prepend_translation(&mut self, translation: Vec2)

Add a translation before this similarity.

Doing so will mean that the translation being added will get transformed by this similarity's rotational and scaling parts.

pub fn append_translation(&mut self, translation: Vec2)

Add a translation after this similarity.

Doing so will mean that the translation being added will not transformed by this similarity's rotational or scaling parts.

pub fn prepend_similarity(&mut self, other: Self)

Prepend transformation by another similarity.

This means that the transformation being applied will take place before this similarity, i.e. both its translation and rotation will be rotated by the other similarity's rotational part, and its translation will be scaled by the other similarity's scaling part.

pub fn append_similarity(&mut self, other: Self)

Append transformation by another similarity.

This means that the transformation being applied will take place after this similarity, i.e. this similarity's translation and rotation will be rotated by the other similarity's rotational part, and this similarity's translation will be scaled by the other similarity's scaling pat.

pub fn inverse(&mut self)

pub fn inversed(self) -> Self

pub fn transform_vec(&self, vec: Vec2) -> Vec2

pub fn into_homogeneous_matrix(self) -> Mat3