Struct ultraviolet::vec::Vec2x4

#[repr(C)]
pub struct Vec2x4 {
    pub x: f32x4,
    pub y: f32x4,
}

A set of two coordinates which may be interpreted as a vector or point in 2d space.

Generally this distinction between a point and vector is more of a pain than it is worth to distinguish on a type level, however when converting to and from homogeneous coordinates it is quite important.

Fields
x: f32x4y: f32x4
Implementations

impl Vec2x4

pub const fn new(x: f32x4, y: f32x4) -> Self

pub const fn broadcast(val: f32x4) -> Self

pub fn unit_x() -> Self

pub fn unit_y() -> Self

pub fn into_homogeneous_point(self) -> Vec3x4

Create a homogeneous 2d point from this vector interpreted as a point, meaning the homogeneous component will start with a value of 1.0.

pub fn into_homogeneous_vector(self) -> Vec3x4

Create a homogeneous 2d vector from this vector, meaning the homogeneous component will always have a value of 0.0.

pub fn from_homogeneous_point(v: Vec3x4) -> Self

Create a 2d point from a homogeneous 2d point, performing division by the homogeneous component. This should not be used for homogeneous 2d vectors, which will have 0 as their homogeneous component.

pub fn from_homogeneous_vector(v: Vec3x4) -> Self

Create a 2d vector from homogeneous 2d vector, which simply discards the homogeneous component.

pub fn dot(&self, other: Vec2x4) -> f32x4

pub fn wedge(&self, other: Vec2x4) -> Bivec2x4

The wedge (aka exterior) product of two vectors.

Note: Sometimes called “cross” product in 2D. Such a product is not well defined in 2 dimensions and is really just shorthand notation for a hacky operation that extends the vectors into 3 dimensions, takes the cross product, then returns only the resulting Z component as a pseudoscalar value. This value is will have the same value as the resulting bivector of the wedge product in 2d (a 2d bivector is also a kind of pseudoscalar value), so you may use this product to calculate the same value.

This operation results in a bivector, which represents the plane parallel to the two vectors, and which has a ‘oriented area’ equal to the parallelogram created by extending the two vectors, oriented such that the positive direction is the one which would move self closer to other.

pub fn geom(&self, other: Vec2x4) -> Rotor2x4

The geometric product of this and another vector, which is defined as the sum of the dot product and the wedge product.

This operation results in a ‘rotor’, named as such as it may define a rotation. The rotor which results from the geometric product will rotate in the plane parallel to the two vectors, by twice the angle between them and in the opposite direction (i.e. it will rotate in the direction that would bring other towards self, and rotate in that direction by twice the angle between them).

pub fn rotate_by(&mut self, rotor: Rotor2x4)

pub fn rotated_by(self, rotor: Rotor2x4) -> Self

pub fn reflected(&self, normal: Vec2x4) -> Self

pub fn mag_sq(&self) -> f32x4

pub fn mag(&self) -> f32x4

pub fn normalize(&mut self)

pub fn normalized(&self) -> Self

pub fn mul_add(&self, mul: Vec2x4, add: Vec2x4) -> Self

pub fn abs(&self) -> Self

pub fn clamp(&mut self, min: Self, max: Self)

pub fn clamped(self, min: Self, max: Self) -> Self

pub fn map<F>(&self, f: F) -> Self where
F: Fn(f32x4) -> f32x4,

pub fn apply<F>(&mut self, f: F) where
F: Fn(f32x4) -> f32x4,

pub fn max_by_component(self, other: Self) -> Self

pub fn min_by_component(self, other: Self) -> Self

pub fn component_max(&self) -> f32x4

pub fn component_min(&self) -> f32x4

pub fn zero() -> Self

pub fn one() -> Self

pub fn xyz(&self) -> Vec3x4

pub fn xyzw(&self) -> Vec4x4

pub fn layout() -> Layout

pub fn as_array(&self) -> &[f32x4; 2]

pub fn as_slice(&self) -> &[f32x4]

pub fn as_byte_slice(&self) -> &[u8]

pub fn as_mut_slice(&mut self) -> &mut [f32x4]

pub fn as_mut_byte_slice(&mut self) -> &mut [u8]

pub const fn as_ptr(&self) -> *const f32x4

Returns a constant unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

pub fn as_mut_ptr(&mut self) -> *mut f32x4

Returns a mutable unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

impl Vec2x4

pub fn new_splat(x: f32, y: f32) -> Self

pub fn splat(vec: Vec2) -> Self

pub fn blend(mask: m32x4, tru: Self, fals: Self) -> Self

Blend two vectors together lanewise using mask as a mask.

This is essentially a bitwise blend operation, such that any point where there is a 1 bit in mask, the output will put the bit from tru, while where there is a 0 bit in mask, the output will put the bit from fals

pub fn refract(&mut self, normal: Self, eta: f32x4)

pub fn refracted(&self, normal: Self, eta: f32x4) -> Self