Struct ultraviolet::vec::Vec3

#[repr(C)]pub struct Vec3 {
    pub x: f32,
    pub y: f32,
    pub z: f32,
}

A set of three coordinates which may be interpreted as a point or vector in 3d space, or as a homogeneous 2d vector or point.

Generally this distinction between a point and vector is more of a pain than it is worth to distinguish on a type level, however when converting to and from homogeneous coordinates it is quite important.

Fields
x: f32y: f32z: f32
Implementations

impl Vec3

pub fn new(x: f32, y: f32, z: f32) -> Self

pub fn broadcast(val: f32) -> Self

pub fn unit_x() -> Self

pub fn unit_y() -> Self

pub fn unit_z() -> Self

pub fn into_homogeneous_point(self) -> Vec4

Create a homogeneous 3d point from this vector interpreted as a point, meaning the homogeneous component will start with a value of 1.0.

pub fn into_homogeneous_vector(self) -> Vec4

Create a homogeneous 3d vector from this vector, meaning the homogeneous component will always have a value of 0.0.

pub fn from_homogeneous_point(v: Vec4) -> Self

Create a 3d point from a homogeneous 3d point, performing division by the homogeneous component. This should not be used for homogeneous 3d vectors, which will have 0 as their homogeneous component.

pub fn from_homogeneous_vector(v: Vec4) -> Self

Create a 3d vector from homogeneous 2d vector, which simply discards the homogeneous component.

pub fn dot(&self, other: Vec3) -> f32

pub fn wedge(&self, other: Vec3) -> Bivec3

The wedge (aka exterior) product of two vectors.

This operation results in a bivector, which represents the plane parallel to the two vectors, and which has a 'oriented area' equal to the parallelogram created by extending the two vectors, oriented such that the positive direction is the one which would move self closer to other.

pub fn geom(&self, other: Vec3) -> Rotor3

The geometric product of this and another vector, which is defined as the sum of the dot product and the wedge product.

This operation results in a 'rotor', named as such as it may define a rotation. The rotor which results from the geometric product will rotate in the plane parallel to the two vectors, by twice the angle between them and in the opposite direction (i.e. it will rotate in the direction that would bring other towards self, and rotate in that direction by twice the angle between them).

pub fn rotate_by(&mut self, rotor: Rotor3)

pub fn rotated_by(self, rotor: Rotor3) -> Self

pub fn cross(&self, other: Vec3) -> Self

pub fn reflect(&mut self, normal: Vec3)

pub fn reflected(&self, normal: Vec3) -> Self

pub fn mag_sq(&self) -> f32

pub fn mag(&self) -> f32

pub fn normalize(&mut self)

pub fn normalized(&self) -> Self

pub fn mul_add(&self, mul: Vec3, add: Vec3) -> Self

pub fn abs(&self) -> Self

pub fn clamp(&mut self, min: Self, max: Self)

pub fn clamped(self, min: Self, max: Self) -> Self

pub fn map<F>(&self, f: F) -> Self where
F: Fn(f32) -> f32,

pub fn apply<F>(&mut self, f: F) where
F: Fn(f32) -> f32,

pub fn max_by_component(self, other: Self) -> Self

pub fn min_by_component(self, other: Self) -> Self

pub fn component_max(&self) -> f32

pub fn component_min(&self) -> f32

pub fn zero() -> Self

pub fn one() -> Self

pub fn xy(&self) -> Vec2

pub fn xyzw(&self) -> Vec4

pub fn layout() -> Layout

pub fn as_array(&self) -> &[f32; 3]

pub fn as_slice(&self) -> &[f32]

pub fn as_byte_slice(&self) -> &[u8]

pub fn as_mut_slice(&mut self) -> &mut [f32]

pub fn as_mut_byte_slice(&mut self) -> &mut [u8]

pub fn as_ptr(&self) -> *const f32

Returns a constant unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

pub fn as_mut_ptr(&mut self) -> *mut f32

Returns a mutable unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

impl Vec3

pub fn refract(&mut self, normal: Self, eta: f32)

pub fn refracted(&self, normal: Self, eta: f32) -> Self