Struct ultraviolet::vec::Vec4x8

#[repr(C)]
pub struct Vec4x8 {
    pub x: f32x8,
    pub y: f32x8,
    pub z: f32x8,
    pub w: f32x8,
}

A set of four coordinates which may be interpreted as a point or vector in 4d space, or as a homogeneous 3d vector or point.

Generally this distinction between a point and vector is more of a pain than it is worth to distinguish on a type level, however when converting to and from homogeneous coordinates it is quite important.

Fields
x: f32x8y: f32x8z: f32x8w: f32x8
Implementations

impl Vec4x8

pub const fn new(x: f32x8, y: f32x8, z: f32x8, w: f32x8) -> Self

pub const fn broadcast(val: f32x8) -> Self

pub fn unit_x() -> Self

pub fn unit_y() -> Self

pub fn unit_z() -> Self

pub fn unit_w() -> Self

pub fn dot(&self, other: Vec4x8) -> f32x8

pub fn reflect(&mut self, normal: Vec4x8)

pub fn reflected(&self, normal: Vec4x8) -> Self

pub fn mag_sq(&self) -> f32x8

pub fn mag(&self) -> f32x8

pub fn normalize(&mut self)

pub fn normalized(&self) -> Self

pub fn normalize_homogeneous_point(&mut self)

Normalize self in-place by interpreting it as a homogeneous point, i.e. scaling the vector to ensure the homogeneous component has length 1.

pub fn normalized_homogeneous_point(&self) -> Self

Normalize self by interpreting it as a homogeneous point, i.e. scaling the vector to ensure the homogeneous component has length 1.

pub fn truncated(&self) -> Vec3x8

Convert self into a Vec3 by simply removing its w component.

pub fn mul_add(&self, mul: Vec4x8, add: Vec4x8) -> Self

pub fn abs(&self) -> Self

pub fn clamp(&mut self, min: Self, max: Self)

pub fn clamped(self, min: Self, max: Self) -> Self

pub fn map<F>(&self, f: F) -> Self where
F: Fn(f32x8) -> f32x8,

pub fn apply<F>(&mut self, f: F) where
F: Fn(f32x8) -> f32x8,

pub fn max_by_component(self, other: Self) -> Self

pub fn min_by_component(self, other: Self) -> Self

pub fn component_max(&self) -> f32x8

pub fn component_min(&self) -> f32x8

pub fn zero() -> Self

pub fn one() -> Self

pub const fn xy(&self) -> Vec2x8

pub const fn xyz(&self) -> Vec3x8

pub fn layout() -> Layout

pub fn as_array(&self) -> &[f32x8; 4]

pub fn as_slice(&self) -> &[f32x8]

pub fn as_byte_slice(&self) -> &[u8]

pub fn as_mut_slice(&mut self) -> &mut [f32x8]

pub fn as_mut_byte_slice(&mut self) -> &mut [u8]

pub const fn as_ptr(&self) -> *const f32x8

Returns a constant unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

pub fn as_mut_ptr(&mut self) -> *mut f32x8

Returns a mutable unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

impl Vec4x8

pub fn new_splat(x: f32, y: f32, z: f32, w: f32) -> Self

pub fn splat(vec: Vec4) -> Self

pub fn blend(mask: m32x8, tru: Self, fals: Self) -> Self

Blend two vectors together lanewise using mask as a mask.

This is essentially a bitwise blend operation, such that any point where there is a 1 bit in mask, the output will put the bit from tru, while where there is a 0 bit in mask, the output will put the bit from fals