Trait ultraviolet::interp::Lerp

pub trait Lerp<T> {
    fn lerp(&self, end: Self, t: T) -> Self;
}

Pure linear interpolation, i.e. (1.0 - t) * self + (t) * end.

For interpolating Rotors with linear interpolation, you almost certainly want to normalize the returned Rotor. For example,

let interpolated_rotor = rotor1.lerp(rotor2, 0.5).normalized();

For most cases (especially where performance is the primary concern, like in animation interpolation for games, this ‘normalized lerp’ or ‘nlerp’ is probably what you want to use. However, there are situations in which you really want the interpolation between two Rotors to be of constant angular velocity. In this case, check out Slerp.

Required methods

fn lerp(&self, end: Self, t: T) -> Self