Module ultraviolet::transform

Dedicated transformation types as the combination of primitives.

Note that you may want to us these types over the corresponding type of homogeneous transformation matrix because they are faster in most operations, especially composition and inverse.

Structs

Isometry2

An Isometry, aka a “rigid body transformation”.

Isometry3

An Isometry, aka a “rigid body transformation”.

Isometry2x4

An Isometry, aka a “rigid body transformation”.

Isometry2x8

An Isometry, aka a “rigid body transformation”.

Isometry3x4

An Isometry, aka a “rigid body transformation”.

Isometry3x8

An Isometry, aka a “rigid body transformation”.

Similarity2

A Similarity, i.e. an Isometry but with an added uniform scaling.

Similarity3

A Similarity, i.e. an Isometry but with an added uniform scaling.

Similarity2x4

A Similarity, i.e. an Isometry but with an added uniform scaling.

Similarity2x8

A Similarity, i.e. an Isometry but with an added uniform scaling.

Similarity3x4

A Similarity, i.e. an Isometry but with an added uniform scaling.

Similarity3x8

A Similarity, i.e. an Isometry but with an added uniform scaling.