Module ultraviolet::transform

Dedicated transformation types as the combination of primitives.

Note that you may want to us these types over the corresponding type of homogeneous transformation matrix because they are faster in most operations, especially composition and inverse.

Structs

Isometry2
An Isometry, aka a "rigid body transformation".

Isometry3
An Isometry, aka a "rigid body transformation".

Similarity2
A Similarity, i.e. an Isometry but with an added uniform scaling.

Similarity3
A Similarity, i.e. an Isometry but with an added uniform scaling.

WIsometry2
An Isometry, aka a "rigid body transformation".

WIsometry3
An Isometry, aka a "rigid body transformation".

WSimilarity2
A Similarity, i.e. an Isometry but with an added uniform scaling.

WSimilarity3
A Similarity, i.e. an Isometry but with an added uniform scaling.